move cmd to inspect #17543
